pyproject.toml
[tool.poetry]
name = "hydra-flash"
version = "0.1.0"
description = ""
authors = ["Victor Joos <victor.joos@uclouvain.be>"]

[tool.poetry.dependencies]
python = "^3.7"
torch = "^1.11.0"
pytorch-lightning = "^1.6.4"
hydra-core = "^1.2.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
flashtrain = 'hydra_flash.train:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
